#6: Refactor
