+++ packages/webapp/src/app/layout.tsx
@@ -1,11 +1,16 @@
 import React, { ReactNode } from 'react';
-import Providers from '@/app/components/providers/Providers';
+import Providers from '@/components/providers/Providers';
 
 export interface RootLayoutProps {
   children: ReactNode;
 }
 
-export default async function RootLayout({ children }: RootLayoutProps) {
+export const metadata = {
+  title: 'Food Composer',
+  description: 'Food Composer App',
+};
+
+export default function RootLayout({ children }: RootLayoutProps) {
   return (
     <html lang="en">
       <body>
